// ExistingCode/src/data-access/api/database/data-contracts.ts
export interface CollarType {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	CollarTypeId: string;
	Description: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	collars: Collar[];
}

export interface CompanyType {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	CompanyTypeId: string;
	Description: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	companys: Company[];
}

export interface HoleStatus {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	HoleStatusId: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	holes: Hole[];
}
// ...
export interface SubTarget {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	SubTarget: string;
	Description?: string;
	SortOrder?: number;
	SubTargetId: string;
	Target: string;
	target: Target;
	collars: Collar[];
	drillPlans: DrillPlan[];
}

export interface Target {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Target: string;
	CentroidEasting?: number;
	CentroidNorthing?: number;
	Description?: string;
	Organization: string;
	Priority?: number;
	Project?: string;
	Radius_m?: number;
	SortOrder?: number;
	TargetCode?: string;
	TargetId: string;
	TargetType?: string;
	Tenement: string;
	organization: Organization;
	tenement: Tenement;
	collars: Collar[];
	drillPatterns: DrillPattern[];
	drillPlans: DrillPlan[];
	subTargets: SubTarget[];
}

export interface Tenement {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Tenement: string;
	AnnExpRent: number;
	CostCd: string;
	DataSource: string;
	/** @format date-time */
	EndDt?: string;
	/** @format date-time */
	GrantDt?: string;
	Holder: string;
	LeaseStatus: string;
	Organization: string;
	/** @format date-time */
	RegDt?: string;
	RegStatus: string;
	/** @format date-time */
	StartDt?: string;
	TenementId: string;
	costCd: CostCodes;
	holder: Company;
	leaseStatus: LeaseStatus;
	regStatus: TenementStatus;
	collars: Collar[];
	drillPlans: DrillPlan[];
	drillPrograms: DrillProgram[];
	sites: Site[];
	targets: Target[];
}
// ...
export interface RowStatus {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: number;
	Description: string;
	IsDefaultInd: boolean;
	RowStatusId: string;
	SortOrder: number;
	collars: Collar[];
	collarCoordinates: CollarCoordinate[];
	coreRecoveryRunLogs: CoreRecoveryRunLog[];
	cycloneCleanings: CycloneCleaning[];
	drillMethods: DrillMethod[];
	fractureCountLogs: FractureCountLog[];
	geologyCombinedLogs: GeologyCombinedLog[];
	magSusLogs: MagSusLog[];
	metaDataLogs: MetaDataLog[];
	ptSamples: PtSample[];
	ptSampleQcs: PtSampleQc[];
	rigSetups: RigSetup[];
	rockMechanicLogs: RockMechanicLog[];
	rockQualityDesignationLogs: RockQualityDesignationLog[];
	samples: Sample[];
	sampleQcs: SampleQc[];
	sampleRegisters: SampleRegister[];
	shearLogs: ShearLog[];
	specificGravityPtLogs: SpecificGravityPtLog[];
	standardSampleQcs: StandardSampleQc[];
	structureLogs: StructureLog[];
	structurePtLogs: StructurePtLog[];
	surveys: Survey[];
	surveyLogs: SurveyLog[];
	xrfSamples: XrfSample[];
	xrfSampleQcs: XrfSampleQc[];
}
// ...
export interface Organization {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Organization: string;
	Description?: string;
	OrganizationId: string;
	SortOrder?: number;
	collars: Collar[];
	coreRecoveryRunLogs: CoreRecoveryRunLog[];
	drillMethods: DrillMethod[];
	drillPatterns: DrillPattern[];
	drillPlans: DrillPlan[];
	drillPrograms: DrillProgram[];
	fractureCountLogs: FractureCountLog[];
	geologyCombinedLogs: GeologyCombinedLog[];
	holes: Hole[];
	metaDataLogs: MetaDataLog[];
	ptSamples: PtSample[];
	ptSampleQcs: PtSampleQc[];
	rigSetups: RigSetup[];
	rockMechanicLogs: RockMechanicLog[];
	rockQualityDesignationLogs: RockQualityDesignationLog[];
	samples: Sample[];
	sampleQcs: SampleQc[];
	shearLogs: ShearLog[];
	specificGravityPtLogs: SpecificGravityPtLog[];
	standardSamples: StandardSample[];
	standardSampleQcs: StandardSampleQc[];
	structureLogs: StructureLog[];
	surveys: Survey[];
	targets: Target[];
	xrfSamples: XrfSample[];
	xrfSampleQcs: XrfSampleQc[];
	xrfStandardSamples: XrfStandardSample[];
}
// ...
export interface Hole {
	ReportIncludeInd?: boolean;
	ValidationStatus?: HoleValidationStatusEnum;
	ValidationErrors?: string | null;
	RowStatus?: HoleRowStatusEnum;
	SupersededById?: string;
	ActiveInd?: boolean;
	rv?: string;
	/** @format date-time */
	CreatedOnDt?: string;
	CreatedBy?: string;
	/** @format date-time */
	ModifiedOnDt?: string;
	ModifiedBy?: string;
	HoleId: string;
	HoleStatus: string;
	Organization: string;
	holeStatus: HoleStatus;
	organization: Organization;
	holeNames: HoleName[];
}

export interface HolePurposeDetail {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	HolePurpose: string;
	HolePurposeDetailId: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	holePurpose: HolePurpose;
	drillPlans: DrillPlan[];
}

export interface HolePurpose {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Category?: string;
	Description: string;
	HolePurposeId: string;
	IsDefaultInd: boolean;
	SortOrder: number;
	collars: Collar[];
	drillPlans: DrillPlan[];
	holePurposeDetails: HolePurposeDetail[];
}

export interface HoleType {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	CoreSize?: string;
	Description: string;
	HoleTypeId: string;
	IsDefaultInd: boolean;
	Method?: string;
	SortOrder: number;
	collars: Collar[];
	drillPlans: DrillPlan[];
}

export interface Phase {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Phase: string;
	Description?: string;
	PhaseId: string;
	Project: string;
	SortOrder?: number;
	collars: Collar[];
	drillPlans: DrillPlan[];
}

export interface Pit {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Pit: string;
	Description?: string;
	PitId: string;
	Project: string;
	SortOrder?: number;
	collars: Collar[];
	drillPlans: DrillPlan[];
}

export interface Prospect {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Prospect: string;
	Description?: string;
	Project: string;
	ProspectId: string;
	SortOrder?: number;
	project: Project;
	collars: Collar[];
	drillPlans: DrillPlan[];
}
// ...
export interface Section {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Code: string;
	Description: string;
	IsDefaultInd: boolean;
	SectionId: string;
	SortOrder: number;
	collars: Collar[];
}
// ...
export interface Collar {
	ReportIncludeInd?: boolean;
	ValidationStatus?: CollarValidationStatusEnum;
	ValidationErrors?: string | null;
	RowStatus?: CollarRowStatusEnum;
	SupersededById?: string;
	ActiveInd?: boolean;
	rv?: string;
	/** @format date-time */
	CreatedOnDt?: string;
	CreatedBy?: string;
	/** @format date-time */
	ModifiedOnDt?: string;
	ModifiedBy?: string;
	CollarId: string;
	ApprovedInd: boolean;
	CasingDepth?: number;
	CollarType: string;
	Comments?: string;
	DataSource: string;
	ExplorationCompany: string;
	/** @format date-time */
	FinishedOnDt?: string;
	HolePurpose: string;
	HolePurposeDetail?: string;
	HoleType: string;
	LoggingEventId: string;
	ModelUseInd: boolean;
	Organization: string;
	OrientationTool?: string;
	ParentCollarId?: string;
	Phase: string;
	Pit: string;
	PreCollarDepth?: number;
	PreCollarId?: string;
	Priority?: number;
	Project: string;
	Prospect: string;
	Redox?: string;
	ResponsiblePerson: string;
	ResponsiblePerson2: string;
	Section: string;
	StartDepth?: number;
	/** @format date-time */
	StartedOnDt?: string;
	SubTarget: string;
	Target: string;
	Tenement: string;
	TotalDepth?: number;
	WaterTableDepth?: number;
	/** @format date-time */
	WaterTableDepthMeasuredOnDt?: string;
	collarType: CollarType;
	explorationCompany: Company;
	holePurpose: HolePurpose;
	holeType: HoleType;
	organization: Organization;
	phase: Phase;
	pit: Pit;
	project: Project;
	prospect: Prospect;
	responsiblePerson: Person;
	responsiblePerson2: Person;
	rowStatus: RowStatus;
	section: Section;
	subTarget: SubTarget;
	target: Target;
	tenement: Tenement;
	collarCoordinates: CollarCoordinate[];
	coreRecoveryRunLogs: CoreRecoveryRunLog[];
	drillMethods: DrillMethod[];
	fractureCountLogs: FractureCountLog[];
	geologyCombinedLogs: GeologyCombinedLog[];
	labDispatchs: LabDispatch[];
	magSusLogs: MagSusLog[];
	metaDataLogs: MetaDataLog[];
	rockMechanicLogs: RockMechanicLog[];
	rockQualityDesignationLogs: RockQualityDesignationLog[];
	samples: Sample[];
	sampleDispatchs: SampleDispatch[];
	sampleQcs: SampleQc[];
	shearLogs: ShearLog[];
	specificGravityPtLogs: SpecificGravityPtLog[];
	structureLogs: StructureLog[];
	surveys: Survey[];
	validationErrors: ValidationError[];
	xrfSamples: XrfSample[];
}

export interface Project {
	ActiveInd: boolean;
	rv: string;
	/** @format date-time */
	CreatedOnDt: string;
	CreatedBy: string;
	/** @format date-time */
	ModifiedOnDt: string;
	ModifiedBy: string;
	Project: string;
	Description?: string;
	Organization: string;
	ProjectId: string;
	SortOrder?: number;
	collars: Collar[];
	drillPlans: DrillPlan[];
	drillPrograms: DrillProgram[];
	filteredsets: Filteredset[];
	prospects: Prospect[];
}
// ...
export interface VwCollar {
	ActiveInd: boolean;
	ApprovedInd: boolean;
	CasingDepth?: number;
	CollarId: string;
	CollarType?: string;
	Comments?: string;
	CreatedBy: string;
	/** @format date-time */
	CreatedOnDt: string;
	DataSource: string;
	DrillHoleId: string;
	ExplorationCompany?: string;
	/** @format date-time */
	FinishedOnDt?: string;
	HoleNm?: string;
	HolePurpose?: string;
	HolePurposeDetail?: string;
	HoleStatus?: string;
	HoleType?: string;
	LoggingEventId: string;
	ModelUseInd: boolean;
	ModifiedBy?: string;
	/** @format date-time */
	ModifiedOnDt?: string;
	Organization: string;
	OrientationTool?: string;
	OtherHoleNm?: string;
	ParentCollarId?: string;
	Phase?: string;
	Pit?: string;
	PlannedHoleNm?: string;
	PlannedTotalDepth?: number;
	PreCollarDepth?: number;
	PreCollarId?: string;
	Priority?: number;
	Project?: string;
	ProposedHoleNm?: string;
	Prospect?: string;
	Redox?: string;
	ReportIncludeInd: boolean;
	ResponsiblePerson?: string;
	ResponsiblePerson2?: string;
	Section?: string;
	StartDepth?: number;
	/** @format date-time */
	StartedOnDt?: string;
	SubTarget?: string;
	SupersededById?: string;
	Target?: string;
	Tenement?: string;
	TotalDepth?: number;
	ValidationErrors?: string;
	ValidationStatus: number;
	vwCollarId: string;
	WaterTableDepth?: number;
	/** @format date-time */
	WaterTableDepthMeasuredOnDt?: string;
	RowStatus?: RowStatus;
	rv?: string;
}